refactor: split realtime event bridging into its own module
- move event serialization, session-event forwarding, client payload
  handling, and the Twilio forwarder out of integrations/realtime.py
  into integrations/realtime_events.py
- re-import the forwarders in realtime.py so
  agency_swarm.integrations.realtime keeps exposing them unchanged
- point the sanitize-history test at the module that now owns the helper
- realtime.py drops from 584 to 369 lines, back under the 500-line limit
